refactor :: components folder structure
- BookBar 에서 BookList로 잘못된 이름 변경
- KeyboardDismiss 컴포넌트를 keyboardDismiss 폴더 하위 파일로 이동
